+++ RELEASENOTES.md
@@ -11,6 +11,9 @@ November 17, 2018
 - Corrected continuation token examples for `Get-CosmosDbDocument`.
 - Updated CI pipeline to use Pester 4.4.2.
 - Updated CI pipeline to use PlatyPS 0.12.
+- Renamed `ResultHeaders` parameter to `ResponseHeader` in
+  `Get-CosmosDbDocuments` function to adhere to PowerShell standards,
+  but included alias for `ResultHeaders` to prevent breaking change.
 
 ## What is New in CosmosDB 2.1.14.220