+++ RELEASENOTES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## What is New in CosmosDB Unreleased
+## What is New in CosmosDB 2.1.15.237
 
 November 17, 2018